+++ admin/add-new-category.php
@@ -1,5 +1,6 @@
 <?php
 include "header.php";
+$rolee = $_SESSION['b2b_admin_role'];
 
 if(isset($_GET['id']))
 {
@@ -27,58 +28,61 @@ else
     <!-- MAIN CONTENT -->
     <div id="content">
         <!-- row -->
-        <div class="row">
 
-            <!-- col -->
-            <div class="col-xs-12 col-sm-12 col-md-12 col-lg-12">
-                <h1 class="page-title txt-color-blueDark"><!-- PAGE HEADER --><i class="fa-fw fa fa-briefcase "></i> Add Category To <?=$restaurant_name?> Restaurant</h1>
+
+        <?php if ($rolee == 1) {?>
+            <div class="row">
+
+                <!-- col -->
+                <div class="col-xs-12 col-sm-12 col-md-12 col-lg-12">
+                    <h1 class="page-title txt-color-blueDark"><!-- PAGE HEADER --><i class="fa-fw fa fa-briefcase "></i> Add Category To <?=$restaurant_name?> Restaurant</h1>
+                </div>
+
             </div>
 
-        </div>
+            <br>
+            <div align="left">
+                <form  action="import-category.php" method="post"  enctype="multipart/form-data">
+                    <fieldset>
 
-        <br>
-        <div align="left">
-            <form  action="import-category.php" method="post"  enctype="multipart/form-data">
-                <fieldset>
+                        <div class="form-group">
+                            <label>Import Category Through CSV</label>
+                            <input class="form-control" id="file" name="file"  type="file" required>
+                            <input type="hidden" value="<?=$menu_id?>" name="menu_id" id="menu_id">
+                            <input type="hidden" value="<?=$_SERVER['REQUEST_URI']?>" name="url" id="url">
 
-                    <div class="form-group">
-                        <label>Import Category Through CSV</label>
-                        <input class="form-control" id="file" name="file"  type="file" required>
-                        <input type="hidden" value="<?=$menu_id?>" name="menu_id" id="menu_id">
-                        <input type="hidden" value="<?=$_SERVER['REQUEST_URI']?>" name="url" id="url">
+                        </div>
 
-                    </div>
+                    </fieldset>
+                    <button name="Import" type="submit" class="btn btn-primary"  data-loading-text="Loading...">
+                        Import CSV File
+                    </button><br>
+                    *Please see the sample CSV file link. <a target="_blank" href="https://docs.google.com/spreadsheets/d/1whfVt9HdEAb2_ldgB03spciFwTG6pdaUxKU-Ocur4ZM/edit?usp=sharing">Click Here</a>
+                </form>
+            </div><br><br>
 
-                </fieldset>
-                <button name="Import" type="submit" class="btn btn-primary"  data-loading-text="Loading...">
-                    Import CSV File
-                </button><br>
-                *Please see the sample CSV file link. <a target="_blank" href="https://docs.google.com/spreadsheets/d/1whfVt9HdEAb2_ldgB03spciFwTG6pdaUxKU-Ocur4ZM/edit?usp=sharing">Click Here</a>
-            </form>
-        </div><br><br>
+            <!-- widget grid -->
+            <?php
 
-        <!-- widget grid -->
-        <?php
-
-        if(!empty($categories)) {
-            $counter = 0;
-            foreach($categories as $cat)
-            {
-                DB::queryFirstRow("select * from items where category_id =  '".$cat['id']."'");
-                if(DB::count() > 0)
+            if(!empty($categories)) {
+                $counter = 0;
+                foreach($categories as $cat)
                 {
-                    $counter = 1;
+                    DB::queryFirstRow("select * from items where category_id =  '".$cat['id']."'");
+                    if(DB::count() > 0)
+                    {
+                        $counter = 1;
+                    }
                 }
-            }
-            if($counter == 1){
-        ?>
-        <div align="center">
-            <a style="text-decoration: none" href="add-business-offers.php?id=<?=$menu_id?>"><div class="btn btn-purple btn-lg">
-                    <i class="fa fa-plus"></i>
-                    Add Business Offers To <?=$restaurant_name?>
-                </div></a>
-        </div><br><br>
-        <?php  } }?>
+                if($counter == 1){
+                    ?>
+                    <div align="center">
+                        <a style="text-decoration: none" href="add-business-offers.php?id=<?=$menu_id?>"><div class="btn btn-purple btn-lg">
+                                <i class="fa fa-plus"></i>
+                                Add Business Offers To <?=$restaurant_name?>
+                            </div></a>
+                    </div><br><br>
+                <?php  } } }?>
         <section id="widget-grid"  id="myform">
 
             <!-- SHOW CATEGORIES-->
@@ -113,9 +117,13 @@ else
                                             <th data-class="expand"><i class="fa fa-fw fa-user text-muted hidden-md hidden-sm hidden-xs"></i> Logo</th>
                                             <th data-hide="phone"><i class="fa-fw fa fa-info text-muted hidden-md hidden-sm hidden-xs"></i> Name </th>
                                             <th data-hide="phone"><i class="fa-fw fa fa-info text-muted hidden-md hidden-sm hidden-xs"></i> שֵׁם </th>
-                                            <th data-hide="phone"><i class="fa-fw fa fa-plus text-muted hidden-md hidden-sm hidden-xs"></i> Add Items </th>
-                                            <th data-hide="phone,tablet"><i class="fa fa-fw fa-edit txt-color-blue hidden-md hidden-sm hidden-xs"></i> Edit</th>
-                                            <th data-hide="phone,tablet"><i class="fa fa-fw fa-edit txt-color-blue hidden-md hidden-sm hidden-xs"></i> Delete</th>
+                                            <th data-hide="phone"><i class="fa-fw fa fa-plus text-muted hidden-md hidden-sm hidden-xs"></i>  Items </th>
+
+
+                                            <?php if ($rolee == 1) {?>
+                                                <th data-hide="phone,tablet"><i class="fa fa-fw fa-edit txt-color-blue hidden-md hidden-sm hidden-xs"></i> Edit</th>
+                                                <th data-hide="phone,tablet"><i class="fa fa-fw fa-edit txt-color-blue hidden-md hidden-sm hidden-xs"></i> Delete</th>
+                                            <?php } ?>
                                         </tr>
                                         </thead>
                                         <tbody>
@@ -131,10 +139,12 @@ else
                                                 <td><?=$category['name_en']?></td>
                                                 <td><?=$category['name_he']?></td>
 
-                                                <td><a style="text-decoration: none" href="add-new-items.php?id=<?=$category['id']?>"><button class="btn btn-labeled btn-success  txt-color-white add" style="border-color: #4c4f53;"><i class="fa fa-fw fa-plus"></i> Add Items </button></a></td>
-                                                <td><a href="edit-category.php?id=<?=$category['id']?>"><button class="btn btn-labeled btn-primary bg-color-blueDark txt-color-white add" style="border-color: #4c4f53;"><i class="fa fa-fw fa-edit"></i> Edit </button></a></td>
-                                                <td><a onclick="delete_category('<?=$category['id']?>','<?=$_SERVER['REQUEST_URI']?>')"><button class="btn btn-labeled btn-danger txt-color-white add" style="border-color: #4c4f53;"><i class="fa fa-fw fa-trash-o"></i> Delete</button></a></td>
+                                                <td><a style="text-decoration: none" href="add-new-items.php?id=<?=$category['id']?>"><button class="btn btn-labeled btn-success  txt-color-white add" style="border-color: #4c4f53;"><i class="fa fa-fw fa-plus"></i> Items </button></a></td>
 
+                                                <?php if ($rolee == 1) {?>
+                                                    <td><a href="edit-category.php?id=<?=$category['id']?>"><button class="btn btn-labeled btn-primary bg-color-blueDark txt-color-white add" style="border-color: #4c4f53;"><i class="fa fa-fw fa-edit"></i> Edit </button></a></td>
+                                                    <td><a onclick="delete_category('<?=$category['id']?>','<?=$_SERVER['REQUEST_URI']?>')"><button class="btn btn-labeled btn-danger txt-color-white add" style="border-color: #4c4f53;"><i class="fa fa-fw fa-trash-o"></i> Delete</button></a></td>
+                                                <?php } ?>
                                             </tr>
                                         <?php  } ?>
 
@@ -165,7 +175,7 @@ else
                     }
                     reader.addEventListener("load", function () {
 
-                       // alert(reader.result);
+                        // alert(reader.result);
                         globalImgCat = reader.result;
 
                     }, false);
@@ -199,10 +209,15 @@ else
                                     <i class="fa fa-arrow-left"></i>
                                     Go Back
                                 </div>
+
+
+                                <?php if ($rolee == 1) {?>
                                 <div onclick="show_category_div()" class="btn btn-primary btn-lg">
                                     <i class="fa fa-plus"></i>
                                     Add Category
                                 </div>
+                                <?php } ?>
+
                                 <br><br>
                                 <div id="add-category" style="display: none">
                                     <form>
